Add support for non .org recommended products.
* ENHANCEMENT: Added logic to the Memberlite Guide page to support products privately hosted (not on .org).

Tweaks were made to the function `memberlite_plugin_action_button` to take a third parameter and pass in a download link. If this download link is present it also opens in a new tab so people may download and install the product.

This keeps uniformity in the functionality of finding the plugin in the plugins.php search in the WordPress dashboard.

Both Paid Memberships Pro and Sitewide Sales are not currently found at all if you use the "Install" button on the block.